Hi Duane,
        Anthony Baxter and myself have completed the patches for comm.c to
get it to use poll(2) directly rather than select().  This provides a largish
performance increase for systems that implement select() as a library call
(such as Solaris), and also allows systems such as Solaris to bypass the silly
1024 open FD limit which library select() has.  We've tested this code on
running systems and found a smallish increase in overall ability to service
connections, and a respectable drop in CPU usage.  To use, just compile with
the -DUSE_POLL flag in the make options.  Patch applies to 1.1b25 version
of comm.c.  As with all code, there probably are still bugs...