From:    Dancer <dancer@zeor.simegen.com>
It turns out that if you are using an external authenticator, the
username is only logged on _subsequent_ (that is, cached) authentication
hits. If the user's authentication credentials expire from the cache,
their next (initially authenticated) request doesn't log any user info.

I'm pretty sure that this behaviour is unintentional.
Here's a patch: