hno squid-2.2.STABLE5.log_ip_on_direct.patch with additions
Squid-2.2.STABLE5: Log destination IP on DIRECT
This patch logs the destination IP as part of the hierarchy tag in
access.log when going direct. This has been requested by a number of
people from accounting reasons, and logging the hostname is mostly
redundant as it is part of the URL as well.

Additions: Can be rewerted to the old behaviour by turning log_ip_on_direct off