hno squid-2.3.DEVEL2.cache_swap_log_name-2.patch Squid-2.3.DEVEL2: cache_swap_log name based on cache_dir name
%s can now be used in cache_swap_log to build a log file name based on
the cache_dir name instead of numbering the files according to their
corresponding cache_dir location in squid.conf. This is very useful if
you'd like to be able to add or remove cache directories while using
cache_swap_log to have the index files stored outside the cache
directories.