From Robert Collins:
The integer types patch introduced a problem with squid.h including
sys/types.h before defining FD_SETSIZE. This may cause problems on
IRIX, NETBSD, and definately does on cygwin ...

attached is a patch against HEAD to alter the location that the
squid_types.h is pulled in. I've done a make clean with this and it
was fine.

The changes to md5.c & h are because it depends on known int types.