Bugzilla #215: NULL pointer access for proxy requests in accel-only mode
This patch fixes the bug by checking for NULL in aclIsProxyAuth().
If access is denied due to receipt of a proxy request with
'httpd_accel_with_proxy off' then AclMatchedName is NULL.
clientAccessCheckDone() calls aclIsProxyAuth() with the NULL pointer,
so the check could go into either function.  It was cleaner to put
it in aclIsProxyAuth().