Security fixes.
These changes are the result of a report by zen-parse@gmx.net regarding
a bug in rfc1035NameUnpack().  When handling "compressed" names, we didn't
check the value of the "pointer" to make sure it points to inside the
reply buffer.

I changed the static unpack function interfaces so that instead of
returning a new offset value they now return success or failure.
The offset is passed as a pointer.

I added a fake DNS error code (15) for use when the reply message
is bogus or unsafe to parse.  This is currently the only indication
of a problem.  The implementation doesn't say why unpacking failed.
This library doesn't have any hooks to Squid's debugging/logging.
We could use syslog I suppose....

I tried to make sure that only programming bugs, not bogus replies,
can cause an assertion.  i.e., some of the former assertions have
been changed to just return an error status.