Bug #490: assertion failed: ssl.cc:228: "len == (size_t)sslState- >client.len" in sslWriteServerDone
comm_write returned the number of bytes written in the last segment if the
write got split in multiple segments, not the total written size.

I also removed some bogous special case for write() == 0 which was
copy-pasted from comm_read. A write() == 0 only occurs if 0 bytes
was requested to be written, and is no different from a normal write