Summary: More MSVC compatability tweaks. Keywords:
More pointer construction to pointer assignment replacements.
Add a test case for delete - checking that the correct delete operator is used in a class hierarchy.
Partly fix g++ 2.95 builds - there sstream includes the system assert, which breaks our custom assert macro. So, a couple of undefs and we have an ugly fix.