Summary: Various fixes - comms and diskio. Keywords:
Comm accept logic was not indicating that it was not cbdata.
Also, mark eventAdd as extern, not C extern.
UFS IO module was not reporting failure on failed opens.
AUFS - log when reads are scheduled, as queued reads are scheduled after the read_ call.
Fix a race condition with synchronous store opens, that previously lead to double-callbacks.