Hi,
On Windows (native and Cygwin) and OS/2, when running
UFSSwapDir::closeTmpSwapLog() in store_dir_ufs.cc, Squid can be fail if the
target file for a rename operation was already deleted.

Because xrename() already try to remove the destination target on Windows,
a previous unlink() is not needed.

This patch extended the native Windows xrename() behaviour to Cygwin and
OS/2 too and remove not needed unlink().

Regards

Guido