Bug #736: ICP dynamic timeout algorithm ignores multicast
The algorithm that calculates the timeout for a set of ICP
queries ignores multicast neighbors.  It also ignores the
expected number of replies because "*exprep" is always set
equal to parent_exprep + sibling_exprep.

This patch adds multicast group counts to the expected number
of ICP replies.  It also sets the query timeout to the
multicast peer rtt (x2) if there were no parents query.  If
there are parents, the expected parents timeout is used.