Bug #1190: Denial of service with forged WCCP messages
WCCP_I_SEE_YOU messages contain a 'number of caches' field which
should be between 1 and 32.  Values outside that range may
crash Squid if WCCP is enabled, and if an attacker can
spoof UDP packets with the WCCP router's IP address.

This patch drops and logs WCCP_I_SEE_YOU messages if the number of
caches is less than 1 or greater than 32.

Forward port of Duane's 2.5 patch.