Bug #321: Squid date handling fails to handle several slightly nonconforming date formats in use on the Internet today.
There is many servers sending non-conforming dates outside the formats
documented in RFC2616. This patch simplifies the date parser allowing
Squid to process dates is a number of other "odd" formats. The benefit
of doing this is that it allows Squid to properly cache objects from
these servers.

Forward port of 2.5 patch.