Bug #1308: squid -k reconfigure internal corruption if the type of a cache_dir is changed
Failed to detect if the type of an existing cache_dir was changed,
calling the parser function of the new type with the internal data of
the existing one..

This patch detects this and logs to cache.log (and the console) that a
restart is required.

Forward port of 2.5 patch.