Support for Method header in ICAP OPTIONS respose.
Currently the 'icap_service' directive in squid.conf specifies both the
URI and method.  The ICAP server also returns the method in the OPTIONS
response.  This could create a conflict if the user specifies the wrong
method for a service.

For now we trust the squid.conf value and issue a warning if the OPTIONS
response reports a different method.