Replacing ClientBody class with BodyReader.
The old ClientBody code did not allow us to insert ICAP into the
flow of an HTTP request.  Code in http.cc called read functions in
client_side.cc when forwarding a request body.  But with ICAP in
the middle, HTTP needs to get the message body from ICAP, not
client_side.

The new BodyReader is similar to ClientBody.  Now read and abort
functions are pointers, rather than hard-coded in HTTP/FTP modules.