Bug #624: Various assertion errors on Range requests
use reply range spec if does not match request range spec (forward as-is)

This assertion is triggered if request range is not satisified by reply.  In
this case, we simply forward the reply as-is.
To forward as-is, we override client_side reply range spec with server_side
reply range spec.


if reply has ranges but request is not: we forward as-is.

Another situation where this assertion is triggered:  if reply is partial
content (status 206) but request is not.  This patch forward as-is the
offending response received on server_side.