Bug #1716: Squid 2.6-STABLE2 does not compile with newer versions of OpenSSL
This patch partially drops the support for OpenSSL 0.9.7x:

When using gcc 3.3.x and OpenSSL 0.9.7x sometimes a compile error can occur.
This is caused by an unpredicatble gcc behaviour on a cast of the first argument
of PEM_ASN1_write(). For this reason this code section is disabled. To enable it,
define ALLOW_ALWAYS_SSL_SESSION_DETAIL=1.

Also ran Astyle.