Fix a bug which resulted in objects being released after forwarding was complete.
This may have been my fault; but the forward code is all a bit weird at
the present time. Bring the forward logic back into line with squid-2.6.
This does need to be verified to be sure its doing the right thing.
It is, at least, resulting in cachable objects once again and hasn't
broken the FTP error page generation.