Forward port of 2.x changes from Duane:
Set errno passed from diskd to squid.

The diskd helper passes errno in responses.  With this patch, errno
is now set in the main squid process after processing a diskd
response.

Initialize unused fields of diomsg in messages from diskd to squid.

Previously, the diskd helper did not initialize the unused fields
of diomsg.  This was annoying when looking at debugging output
and trying to match requests with responses.  This patch initializes
the unused fields.