This change should fix bug #1898: "assertion failed: tunnel.cc:372"
The assertion fails in SslStateData::Connection::dataSent method which is
called by SslStateData::writeClientDone and SslStateData::writeServerDone
methods.

Both write*Done methods did not check for the comm_err_t flag.

When the SSL side closes the connection after a successful but partial write,
write*Done is called with the len argument equal to the number of written
bytes, which is positive but less than the size of the given data for write,
and the dataSent() assertion "amount == (size_t)len" fails.

Based on Christos Tsantilas' work.