Fixed SEGV during reconfigure due to NULL pointer bug in eventDelete()
Since the cancel method may now delete multiple events (when
arg is NULL) it no longer returns after a deletion and
we have a potential NULL pointer problem.  If we just
deleted the last event in the list then *E is now equal
to NULL.  We need to break here or else we'll get a NULL
pointer dereference in the last clause of the for loop.