Bug #1630: TPROXY needs porting from 2.6 to 3.x
This patch adds Linux TPROXY support to Squid 3 forward
porting 2.6 code. Based on Steven Swilton work.

The changes are verified to not break non-TPROXY use, while the
TPROXY code is still untested.
If there will be issues in the optional TPROXY code we deal with
them as bugs later on.