Author: Amos Jeffries <squid3@treenet.co.nz>
Remove cppunit from the Squid distribution

Packaging of cppunit within the squid sources has caused more trouble
than was solved. Ease of install independently of squid combined with
the apparently widespread use of cppunit as a standard tool we feel
there is no need for it to be included.

Maintainers and source testers will need to have cppunit installed on
their systems or not use "make check".