Change diskd default DebugLevel to 0
At DebugLevel = 1, diskd prints "errors" to cache.log when operations
such as open, read, and unlink fail.  These "errors" are not fatal.
In fact, some are to be expected due to the asynchronous nature of
diskd.  The upper layers of Squid are designed to handle these
"errors" gracefully.  They should not be in the log by default
because naive users may report them as bugs.