Some debugging logs seem to indicate that asynchronous calls are fired out of order. The code cannot (and probably should not) handle that.
This out-of-order calls may be happening because async calls are implemented
using time-based events, and events are using wall clock (current_dtime)
when looking for the right place in the event queue. The wall clock may
go backwards, placing a later call earlier in the event queue.

This change forces async call-looking events to use absolute timestamp of
zero, so nothing (including a later async call) can get in front of them.