Author: Henrik Nordstrom <henrik@henriknordstrom.net> Bug #1872: Date parsing error causing HTTP objects to get unexpectedly cached and fresh ICAP OPTIONS rejected.
Applied Henrik's patch from Squid2 patchset #11161. The bug was affecting ICAP
Options TTLs, marking some ICAP services with soon-to-expire options as being
down. Original patch text is below.

An unexpected and undesired sideeffect of the earlier change to accept dates
far in the future was that the date calculations suddenly got daylight savings
sensitive due to mktime() automatically adjusting the supplied date. This has the
unfortunate effect that if Squid is running in a timezone with daylight savings
things may get cached longer than they should.