Partial bug #2034 fix: Inform BodyPipe of the body size when known.
This fix helps the ICAP client to detect aborts and to announce the end of the
HTTP message to the ICAP server.

A bigger problem is that the HTTP side is using the adapted response to make
decisions about the origin server connection. That bug still needs to be
fixed.