Fix memory leaks in NTLM session helper RFCNB_Call(...)
Coverity checker identified that the connect structures were not being
free'd on any of the error cases.

This patch adds several calls to free(con) to prevent the leaks.
It also adds a call to close any partial connections properly when
RFCNB_Session_Req fails. Previously it would leave the open TCP
connection hanging and return a NULL pointer.