Make tcp_recv_bufsize apply to send buffers as well, in an attempt to control window scaling
the main reason to set tcp_recv_bufsize is to avoid problems related to large
windows. For example many firewalls still have problems with window scaling.

this change makes the option apply to send buffers as well, and also hints
Linux that Squid will never increase the buffer so the window scaling should
be clamped to the indicated buffer size.