Bug #2116 fix: Better handling of HTTP 206 Partial Content responses.
When Squid is not doing the ranges it may still forward a Range request to the
origin server and receive a range response. The old code was comparing
endOffset of memObject with the last expected offset according to the Range
header value. That comparison did not account for the header size in
endOffset.

I do not know if the header size should not have been there or
should have been accounted for in the comparison. Adding headers size made
the 206 problem more difficult to reproduce (but it was still there).

Instead of trying to figure out when the header should or should not be
counted, I used http->out.offset and reply->content_range->spec.length.  That
seems to work. This change, however, is in my Puzzle Area of Squid; I have low
confidence the fix is correct and cannot be improved.