Fix-fix for MD5.
Pre-compiler only tested on Linux and FreeBSD.

This update amends the previous to allow OS which provide a partial MD5
implementation but do not supply correct buffer size (MD5_DIGEST_* macro)
to build using the squid bundled code.

To evade symbol-clashes the squid code is also updated to use xMD5Init,
xMD5Update, xMD5Final and the code sorts out which version is to be used
at compile time from configure options and available sources.

For MacOS X and other broken OS the sys/types.h must also be included on
behalf of the sys/*.h which need it.