Bug 2096: allow pending cache hits when delay pools not compiled in
For some time now, Squid has marked pending cache hits (a cache hit
when e->store_status != STORE_OK) as LOG_TCP_MISS.  This was done
so that pending hits do not bypass delay pools.  It had the negative
side effect of generating confusing access.log entries because cache
hits get logged as miss.

With this patch, the behavior is reverted when delay pools are not
compiled in.  Pending hits will be logged as some kind of TCP_HIT.

Users of delay pools will just have to live with this logging quirk,
although the "forwarded to" field of access.log can always be used
to see if the request was forwarded (a miss) or not (a hit).