Formal Debug Levels part 1 - Define names for some debug levels.
This allows the use of DBG_CRITICAL and DBG_IMPORTANT instead of magic
numbers 0 and 1 in the debugs() calls. Making code a little more readable
and easier to use. The code itself is not modified to use them yet.

DBG_DATA is added for level 9, though that may change.

TODO: the other levels (2-8) still need deciding. No consensus was reached
amongst the developers for those levels when discussed.