Moved configuration options that do not depend on the adaptation method from ICAP/ICAPConfig to adaptation/Config. ICAP and eCAP will not share the same base configuration but will share the same base configuration code (i.e., there will be two independent ICAPConfig and ecap::Config objects, but they will have a common parent).
Implemented delayed creation of adaptation services. We used to create
ICAPServiceRep objects when parsing the configuration file.
Create-as-you-parse is imperfect for several reasons, especially if the
services are dynamically loaded as is the case with eCAP. We now remember the
service configuration and then create the actual service object _after_ the
configuration has been parsed and loadable modules, if any, have been loaded.

No functional changes are expected from this change.