Fixed and polished autoconsumption mode.
Fixed: We need to start autoconsuming when new data is appended and we
have not started (but enabled autoconsumpiton) before.

Polished: When notifying a consumer, checking whether mustAutoConsume is
set is pointless as it has no effect on consumer (if any). This check was
probably a leftover from pre-BodySink days.