Added Adaptation::Message that brings together the HttpMsg and the corresponding body_pipe and handle HttpMsg locking.
TODO: This class should not be needed. HttpMsg already has body_pipe. In the
past, we could not use that pipe "as is" because some other code would steal
it from the message. Once such cases are removed and something like
HttpMsg::Pointer is added to handle locking, there will be no need for
Adaptation::Message.

TODO: ICAPInOut should be replaced with two Adaptation::Messages, one for the
virgin message (that often has a body) and one for the cause, even though the
ICAP cause does not need a body.