Bug 2223: Feature: flexible handling of x-forwarded-for
This patch adds three settings for the 'forwarded_for' option in squid.conf:

If set to "transparent", Squid will not alter the X-Forwarded-For header in any
way.

If set to "delete", Squid will delete the entire X-Forwarded-For header.

If set to "truncate", Squid will remove all existing X-Forwarded-For entries,
and place itself as the sole entry.

The old options 'on' and 'off' have been left unaltered.