TestBed: Add master test scripts and some control files
This adds the ./test-builds.sh script which runs over an optional OS
specific control build and a set of layered configurations.

 Usage:  bash ./test-builds.sh [os-name]
 Currently supported OS:  debian, ubuntu

A sub-level build script is also added as the main engine of this system.
It takes a pair of control files and attempts a clean build of squid.
Logging as much of the build output as possible for manual oversight.

The component control files are a set of files providing configure options
and environment settings for the build. Their design and use is detailed in:
  http://wiki.squid-cache.org/Features/CodeTestBed