Generate daily langpack tarball
Following the daily dist build this is supposed to bundle all existing
translations into a langpack tarball for distribution.

Squid releases are bundled with their own most-current translations as at
release date. Daily snapshots included.

This bundling is only intended to be built for the HEAD release as a
a drop-in replacement for any error pages bundled per-release with
Squid 2.6 and later.

TODO: auto-translation patch still to merge will generate current
translations for this to work with. At present it only has the old legacy
folders, but should still operate correctly on those.