Author: Francesco Salvestrini and Dustin J. Mitchell Extend configure with ax_with_prog
Locates an installed program binary, placing the result in the precious
variable VARIABLE. Accepts a present VARIABLE, then --with-program, and
failing that searches for program in the given path (which defaults to
the system path). If program is found, VARIABLE is set to the full path
of the binary; if it is not found VARIABLE is set to VALUE-IF-NOT-FOUND
if provided, unchanged otherwise.

NOTE: This macro is based upon the original AX_WITH_PYTHON macro from
      Dustin J. Mitchell <dustin@cs.uchicago.edu>.

COPYLEFT
   Copyright (c) 2008 Francesco Salvestrini <salvestrini@users.sourceforge.net>
   Copyright (c) 2008 Dustin J. Mitchell <dustin@cs.uchicago.edu>